Bug fix: when generating a custom class with a corresponding interface, only create the interface if it doesn't exist.

+++ System/EamonDD/Game/Menus/ActionMenus/UtilitiesMenus/AdventureSupportMenus/AdventureSupportMenu01.cs
@@ -673,13 +673,16 @@ namespace YourAdventureName.YourGameNamespaceName
 					{
 						var childInterfaceFileName = childClassFileName.Replace(@"\Game\", @"\Framework\").Replace(@"\" + yourClassName + ".cs", @"\" + yourInterfaceName + ".cs");
 
-						var childInterfacePath = Globals.Path.GetDirectoryName(childInterfaceFileName);
+						if (!Globals.File.Exists(childInterfaceFileName))
+						{
+							var childInterfacePath = Globals.Path.GetDirectoryName(childInterfaceFileName);
 
-						Globals.Directory.CreateDirectory(childInterfacePath);
+							Globals.Directory.CreateDirectory(childInterfacePath);
 
-						fileText = InterfaceCsText.Replace("EamonLibraryName", eamonLibraryName).Replace("YourFrameworkNamespaceName", yourFrameworkNamespaceName).Replace("YourInterfaceName", yourInterfaceName);
+							fileText = InterfaceCsText.Replace("EamonLibraryName", eamonLibraryName).Replace("YourFrameworkNamespaceName", yourFrameworkNamespaceName).Replace("YourInterfaceName", yourInterfaceName);
 
-						Globals.File.WriteAllText(childInterfaceFileName, ReplaceMacros(fileText));
+							Globals.File.WriteAllText(childInterfaceFileName, ReplaceMacros(fileText));
+						}
 					}
 
 					var yourEamonUsingStatement = string.Empty;
